+++ api-service/service/frontend_api_test.go
@@ -79,7 +79,7 @@ func TestGetInfoForFrontendDashboard_Success(t *testing.T) {
 		},
 	}
 
-	getTechnologiesUsageRes := map[string]float32{
+	getTechnologiesUsageRes := map[string]float64{
 		"Oracle/Database_HostsCount": 8,
 		"Oracle/Exadata":             0,
 	}
@@ -152,7 +152,7 @@ func TestGetInfoForFrontendDashboard_Success(t *testing.T) {
 		nil,
 	)
 
-	getTechnologiesUsageRes2 := map[string]float32{
+	getTechnologiesUsageRes2 := map[string]float64{
 		"Oracle/Database_HostsCount": 8,
 		"Oracle/Exadata":             2,
 	}
@@ -191,7 +191,7 @@ func TestGetInfoForFrontendDashboard_Fail2(t *testing.T) {
 		Database: db,
 	}
 
-	getTechnologiesUsageRes := map[string]float32{
+	getTechnologiesUsageRes := map[string]float64{
 		"Oracle/Database": 8,
 		"Oracle/Exadata":  0,
 	}
